Added Furigana translation and Japanese text translation

+++ app/src/main/java/com/example/romajijp/model/LrClibTrack.kt
@@ -7,7 +7,7 @@ data class LrclibTrack(
     @SerializedName("trackName")    val trackName: String,
     @SerializedName("artistName")   val artistName: String,
     @SerializedName("albumName")    val albumName: String?,
-    @SerializedName("duration")     val duration: Int,
+    @SerializedName("duration")     val duration: Double,
     @SerializedName("plainLyrics")  val plainLyrics: String?,
     @SerializedName("syncedLyrics") val syncedLyrics: String?,
     @SerializedName("instrumental") val instrumental: Boolean
